+++ src/main/kotlin/io/kotest/plugin/intellij/psi/classes.kt
@@ -4,7 +4,6 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.PsiTreeUtil
 import org.jetbrains.kotlin.asJava.classes.KtLightClass
-import org.jetbrains.kotlin.idea.caches.resolve.analyze
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.KtClass
 import org.jetbrains.kotlin.psi.KtClassOrObject
@@ -12,9 +11,6 @@ import org.jetbrains.kotlin.psi.KtObjectDeclaration
 import org.jetbrains.kotlin.psi.psiUtil.getChildrenOfType
 import org.jetbrains.kotlin.psi.psiUtil.getStrictParentOfType
 import org.jetbrains.kotlin.psi.psiUtil.isAbstract
-import org.jetbrains.kotlin.resolve.BindingContext
-import org.jetbrains.kotlin.resolve.descriptorUtil.classId
-import org.jetbrains.kotlin.types.typeUtil.supertypes
 
 /**
  * Returns the [KtClass] from this light class, otherwise null.
@@ -45,34 +41,6 @@ fun PsiElement.enclosingKtClass(): KtClass? = getStrictParentOfType()
 
 fun PsiElement.enclosingKtClassOrObject(): KtClassOrObject? =
    PsiTreeUtil.getParentOfType(this, KtClassOrObject::class.java)
-
-/**
- * Recursively returns the list of classes and interfaces extended or implemented by the class.
- */
-fun KtClassOrObject.getAllSuperClasses(): List<FqName> {
-   return superTypeListEntries
-      .mapNotNull { it.typeReference }
-      .mapNotNull {
-         runCatching {
-            val bindingContext = it.analyze()
-            bindingContext.get(BindingContext.TYPE, it)
-         }.getOrNull()
-      }.flatMap {
-         runCatching {
-            it.supertypes() + it
-         }.getOrElse { emptyList() }
-      }.mapNotNull {
-         runCatching {
-            it.constructor.declarationDescriptor.classId
-         }.getOrNull()
-      }.mapNotNull {
-         runCatching {
-            val packageName = it.packageFqName
-            val simpleName = it.relativeClassName
-            FqName("$packageName.$simpleName")
-         }.getOrNull()
-      }.filterNot { it.toString() == "kotlin.Any" }
-}
 
 /**
  * Returns true if this [KtClassOrObject] points to a runnable spec object.
