+++ src/main/kotlin/io/kotest/plugin/intellij/toolwindow/KotestTestExplorerService.kt
@@ -17,7 +17,6 @@ import io.kotest.plugin.intellij.psi.specs
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import org.jetbrains.kotlin.idea.core.util.toPsiFile
 import org.jetbrains.kotlin.psi.KtClassOrObject
 import org.jetbrains.kotlin.psi.KtProperty
