+++ src/main/resources/META-INF/plugin.xml
@@ -111,4 +111,8 @@
       </intentionAction>
    </extensions>
 
+   <extensions defaultExtensionNs="org.jetbrains.kotlin">
+      <supportsKotlinPluginMode supportsK2="true" />
+   </extensions>
+
 </idea-plugin>
