+++ src/test/kotlin/io/kotest/plugin/intellij/CallbacksTest.kt
@@ -1,5 +1,6 @@
 package io.kotest.plugin.intellij
 
+import com.intellij.openapi.application.runReadAction
 import com.intellij.testFramework.fixtures.LightJavaCodeInsightFixtureTestCase
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
@@ -16,6 +17,10 @@ class CallbacksTest : LightJavaCodeInsightFixtureTestCase() {
       return path.toString()
    }
 
+   override fun runInDispatchThread(): Boolean {
+      return false
+   }
+
    fun testCallbacks() {
 
       val psiFile = myFixture.configureByFiles(
@@ -23,39 +28,41 @@ class CallbacksTest : LightJavaCodeInsightFixtureTestCase() {
          "/io/kotest/core/spec/style/specs.kt"
       )[0]
 
-      val ktclass = psiFile.specs()[0]
-      val callbacks = ktclass.callbacks()
+      runReadAction {
+         val ktclass = psiFile.specs()[0]
+         val callbacks = ktclass.callbacks()
 
-      callbacks.shouldHaveSize(10)
+         callbacks.shouldHaveSize(10)
 
-      callbacks[0].psi.startOffset shouldBe 115
-      callbacks[0].type shouldBe CallbackType.BeforeTest
+         callbacks[0].psi.startOffset shouldBe 115
+         callbacks[0].type shouldBe CallbackType.BeforeTest
 
-      callbacks[1].psi.startOffset shouldBe 138
-      callbacks[1].type shouldBe CallbackType.AfterTest
+         callbacks[1].psi.startOffset shouldBe 138
+         callbacks[1].type shouldBe CallbackType.AfterTest
 
-      callbacks[2].psi.startOffset shouldBe 160
-      callbacks[2].type shouldBe CallbackType.BeforeSpec
+         callbacks[2].psi.startOffset shouldBe 160
+         callbacks[2].type shouldBe CallbackType.BeforeSpec
 
-      callbacks[3].psi.startOffset shouldBe 183
-      callbacks[3].type shouldBe CallbackType.AfterSpec
+         callbacks[3].psi.startOffset shouldBe 183
+         callbacks[3].type shouldBe CallbackType.AfterSpec
 
-      callbacks[4].psi.startOffset shouldBe 225
-      callbacks[4].type shouldBe CallbackType.AfterEach
+         callbacks[4].psi.startOffset shouldBe 225
+         callbacks[4].type shouldBe CallbackType.AfterEach
 
-      callbacks[5].psi.startOffset shouldBe 243
-      callbacks[5].type shouldBe CallbackType.BeforeEach
+         callbacks[5].psi.startOffset shouldBe 243
+         callbacks[5].type shouldBe CallbackType.BeforeEach
 
-      callbacks[6].psi.startOffset shouldBe 261
-      callbacks[6].type shouldBe CallbackType.AfterContainer
+         callbacks[6].psi.startOffset shouldBe 261
+         callbacks[6].type shouldBe CallbackType.AfterContainer
 
-      callbacks[7].psi.startOffset shouldBe 288
-      callbacks[7].type shouldBe CallbackType.BeforeContainer
+         callbacks[7].psi.startOffset shouldBe 288
+         callbacks[7].type shouldBe CallbackType.BeforeContainer
 
-      callbacks[8].psi.startOffset shouldBe 311
-      callbacks[8].type shouldBe CallbackType.BeforeAny
+         callbacks[8].psi.startOffset shouldBe 311
+         callbacks[8].type shouldBe CallbackType.BeforeAny
 
-      callbacks[9].psi.startOffset shouldBe 328
-      callbacks[9].type shouldBe CallbackType.AfterAny
+         callbacks[9].psi.startOffset shouldBe 328
+         callbacks[9].type shouldBe CallbackType.AfterAny
+      }
    }
 }
